+++ CHANGELOG.rst
@@ -11,6 +11,9 @@ Bug fixes
 * Restore summary in digest email when reporting verbose messages
   (`#278 <https://github.com/fedora-infra/fmn/pull/278>`_).
 
+* X-Fedmsg-Username headers should not be duplicated
+  (`#281 <https://github.com/fedora-infra/fmn/pull/281>`_).
+
 
 v2.0.2
 ======